+++ Classes/Crust.vb
@@ -6,17 +6,18 @@
 '@version: 1.0
 '
 'Description
-' This Crust Class inherits from the Item Class.
+' This 'Crust' Class inherits from the 'Item' Class. Describes crust on a pizza.
+
 Imports EnumLists
 
 Public Class Crust
     Inherits Item
 
     ' CONSTANTS
-    Public Const PAN_PRICE As Decimal = 0.6
-    Public Const REGULAR_PRICE As Decimal = 0.5
-    Public Const THIN_PRICE As Decimal = 0.4
-    Public Const WHEAT_PRICE As Decimal = 0.7
+    Public Const PAN_PRICE As Double = 0.6
+    Public Const REGULAR_PRICE As Double = 0.5
+    Public Const THIN_PRICE As Double = 0.4
+    Public Const WHEAT_PRICE As Double = 0.7
 
     ' instance property
     Public Property CrType As CrustTypeList
@@ -28,7 +29,7 @@ Public Class Crust
     End Sub
 
     ' constructor crust parameters price, crust type
-    Public Sub New(price As Decimal, type As CrustTypeList)
+    Public Sub New(price As Double, type As CrustTypeList)
         MyBase.New(price)
         CrType = type
     End Sub
