+++ Classes/Sauce.vb
@@ -6,18 +6,18 @@
 '@version: 1.0
 '
 'Description
-' This Sauce Class inherits from the Item Class.
+' This 'Sauce' Class inherits from the 'Item' Class. Describes sauce on a pizza.
 Imports EnumLists
 
 Public Class Sauce
     Inherits Item
 
     ' CONSTANTS
-    Public Const TRADITIONAL_PRICE As Decimal = 0.1
-    Public Const ZESTYSALSA_PRICE As Decimal = 0.15
-    Public Const BUFFALO_PRICE As Decimal = 0.2
-    Public Const BBQ_PRICE As Decimal = 0.25
-    Public Const WHITE_PRICE As Decimal = 0.3
+    Public Const TRADITIONAL_PRICE As Double = 0.1
+    Public Const ZESTYSALSA_PRICE As Double = 0.15
+    Public Const BUFFALO_PRICE As Double = 0.2
+    Public Const BBQ_PRICE As Double = 0.25
+    Public Const WHITE_PRICE As Double = 0.3
 
     ' instance property
     Public Property SType As SauceTypeList
@@ -29,7 +29,7 @@ Public Class Sauce
     End Sub
 
     ' constructor sauce parameters price, sauce type
-    Public Sub New(price As Decimal, type As SauceTypeList)
+    Public Sub New(price As Double, type As SauceTypeList)
         MyBase.New(price)
         SType = type
     End Sub
