+++ Classes/Topping.vb
@@ -6,30 +6,30 @@
 '@version: 1.0
 '
 'Description
-' This Topping Class inherits from the Item Class.
+' This 'Topping' Class inherits from the 'Item' Class.
 Imports EnumLists
 
 Public Class Topping
     Inherits Item
 
     ' CONSTANTS
-    Public Const PEPPERONI_PRICE As Decimal = 1.0
-    Public Const HAM_PRICE As Decimal = 1.0
-    Public Const SAUSAGE_PRICE As Decimal = 1.0
-    Public Const BEEF_PRICE As Decimal = 1.0
-    Public Const CHICKEN_PRICE As Decimal = 1.1
-    Public Const BACON_PRICE As Decimal = 1.0
-    Public Const ONION_PRICE As Decimal = 1.0
-    Public Const BELLPEPPER_PRICE As Decimal = 1.0
-    Public Const TOMATO_PRICE As Decimal = 1.1
-    Public Const ARTICHOKE_PRICE As Decimal = 1.2
-    Public Const MUSHROOM_PRICE As Decimal = 1.0
-    Public Const OLIVE_PRICE As Decimal = 1.0
-    Public Const LETTUCE_PRICE As Decimal = 1.2
-    Public Const JALAPENO_PRICE As Decimal = 1.0
-    Public Const BANANAPEPPER_PRICE As Decimal = 1.0
-    Public Const PINEAPPLE_PRICE As Decimal = 1.1
-    Public Const ANCHOVIE_PRICE As Decimal = 1.3
+    Public Const PEPPERONI_PRICE As Double = 1.0
+    Public Const HAM_PRICE As Double = 1.0
+    Public Const SAUSAGE_PRICE As Double = 1.0
+    Public Const BEEF_PRICE As Double = 1.0
+    Public Const CHICKEN_PRICE As Double = 1.1
+    Public Const BACON_PRICE As Double = 1.0
+    Public Const ONION_PRICE As Double = 1.0
+    Public Const BELLPEPPER_PRICE As Double = 1.0
+    Public Const TOMATO_PRICE As Double = 1.1
+    Public Const ARTICHOKE_PRICE As Double = 1.2
+    Public Const MUSHROOM_PRICE As Double = 1.0
+    Public Const OLIVE_PRICE As Double = 1.0
+    Public Const LETTUCE_PRICE As Double = 1.2
+    Public Const JALAPENO_PRICE As Double = 1.0
+    Public Const BANANAPEPPER_PRICE As Double = 1.0
+    Public Const PINEAPPLE_PRICE As Double = 1.1
+    Public Const ANCHOVIE_PRICE As Double = 1.3
 
     ' instance property
     Public Property TType As ToppingTypeList
@@ -41,7 +41,7 @@ Public Class Topping
     End Sub
 
     ' constructor topping parameters price, topping type
-    Public Sub New(price As Decimal, type As ToppingTypeList)
+    Public Sub New(price As Double, type As ToppingTypeList)
         MyBase.New(price)
         TType = type
     End Sub
